chore: refactor timeline out into separate component
